+++ src/card/sets/y2003/mod.rs
@@ -1,0 +1,5 @@
+//! Card sets first released in 2003.
+
+pub(super) mod scourge;
+
+pub(super) use super::{CardRecord, PrintingRecord};